feat: Footer w/Glass 공통 컴포넌트 분리
