+++ apps/ceos/src/pages/activity/index.tsx
@@ -10,9 +10,7 @@ import { Title } from '@ceos/components/Title';
 import { QueryClient, dehydrate, useQuery } from '@tanstack/react-query';
 import { activityApi } from '@ceos-fe/utils';
 import { ResponseInterface } from '@ceos-fe/utils';
-import styled from '@emotion/styled';
-import { GlassShortcutwithTitle } from '@ceos/components/Shortcut';
-import Link from 'next/link';
+import Footer from '@ceos/components/Footer';
 
 // TODO: interface 재정의
 interface ActivityResponse {
@@ -37,6 +35,17 @@ const Activity = () => {
   });
 
   const acitivityList = data?.activityData.data.activities;
+
+  const leftBtn = {
+    title: '더 궁금한 것이 있다면',
+    content: ['자주 묻는 질문', '보러가기'],
+    link: '/FAQ',
+  };
+  const rightBtn = {
+    title: 'CEOS에 참여하고 싶다면',
+    content: ['CEOS 18기', '지원하기'],
+    link: '/recruit',
+  };
 
   return (
     <>
@@ -71,28 +80,7 @@ const Activity = () => {
               );
             })}
           </Flex>
-
-          <RelativeContainer height={500}>
-            <Background src="/shortcuts.svg" />
-            <GlassFlex direction="column" webGap={80}>
-              <Flex webGap={24}>
-                <CustomLink href="/FAQ">
-                  <GlassShortcutwithTitle title="더 궁금한 것이 있다면">
-                    자주 묻는 질문 <br /> 보러가기
-                  </GlassShortcutwithTitle>
-                </CustomLink>
-
-                <CustomLink href="/recruit">
-                  <GlassShortcutwithTitle title="CEOS에 참여하고 싶다면">
-                    CEOS 18기 <br /> 지원하기
-                  </GlassShortcutwithTitle>
-                </CustomLink>
-              </Flex>
-              <Text paletteColor="White" webTypo="Label3">
-                © 2016-2023 CEOS ALL RIGHTS RESERVED.
-              </Text>
-            </GlassFlex>
-          </RelativeContainer>
+          <Footer leftBtn={leftBtn} rightBtn={rightBtn} />
         </Flex>
       </Desktop>
 
@@ -112,27 +100,7 @@ const Activity = () => {
               );
             })}
           </Flex>
-          <RelativeContainer align="start" height={500}>
-            <Background src="/mobileShortcuts.svg" />
-            <GlassFlex direction="column" mobileGap={63}>
-              <Flex mobileGap={17} direction="column">
-                <CustomLink href="/FAQ">
-                  <GlassShortcutwithTitle title="더 궁금한 것이 있다면">
-                    자주 묻는 질문 <br /> 보러가기
-                  </GlassShortcutwithTitle>
-                </CustomLink>
-
-                <CustomLink href="/recruit">
-                  <GlassShortcutwithTitle title="CEOS에 참여하고 싶다면">
-                    CEOS 18기 <br /> 지원하기
-                  </GlassShortcutwithTitle>
-                </CustomLink>
-              </Flex>
-              <Text paletteColor="White" webTypo="Label3">
-                © 2016-2023 CEOS ALL RIGHTS RESERVED.
-              </Text>
-            </GlassFlex>
-          </RelativeContainer>
+          <Footer leftBtn={leftBtn} rightBtn={rightBtn} />
         </Flex>
       </Mobile>
     </>
@@ -158,30 +126,3 @@ export const getStaticProps = async () => {
 };
 
 export default Activity;
-
-const GlassFlex = styled(Flex)`
-  position: absolute;
-  bottom: 80px;
-  z-index: 99;
-  @media (max-width: 1023px) {
-    bottom: 30px;
-  }
-`;
-
-const Background = styled.img`
-  width: 100%;
-  z-index: -99;
-  max-height: 500px;
-
-  @media (max-width: 1023px) {
-    position: absolute;
-    bottom: 0;
-    width: 100vw;
-    max-height: 500px;
-  }
-`;
-
-const CustomLink = styled(Link)`
-  text-decoration: none;
-  color: inherit;
-`;
